+++ src/pkg/backend/clouds/bgcp/instances.go
@@ -39,13 +39,13 @@ import (
 )
 
 type InstanceDetail struct {
-	FirewallTags     []string `yaml:"firewallTags" json:"firewallTags"`
-	TagFingerprint   string   `yaml:"tagFingerprint" json:"tagFingerprint"`
-	LabelFingerprint string   `yaml:"labelFingerprint" json:"labelFingerprint"`
-	Volumes          []instanceVolume
-	FirewallList     backends.FirewallList
-	Network          *backends.Network
-	Subnet           *backends.Subnet
+	FirewallTags     []string         `yaml:"firewallTags" json:"firewallTags"`
+	TagFingerprint   string           `yaml:"tagFingerprint" json:"tagFingerprint"`
+	LabelFingerprint string           `yaml:"labelFingerprint" json:"labelFingerprint"`
+	Volumes          []instanceVolume `yaml:"volumes" json:"volumes"`
+	FirewallIDs      []string         `yaml:"firewallIDs" json:"firewallIDs"`
+	NetworkID        string           `yaml:"networkID" json:"networkID"`
+	SubnetID         string           `yaml:"subnetID" json:"subnetID"`
 }
 
 type instanceVolume struct {
@@ -130,13 +130,6 @@ func (s *b) getInstanceDetails(log *logger.Logger, inst *computepb.Instance, vol
 			sub = ssub[0]
 		}
 	}
-	fwPointers := backends.FirewallList{}
-	for _, fw := range firewalls {
-		fwx := firewallList.WithFirewallID(fw)
-		if fwx.Count() > 0 {
-			fwPointers = append(fwPointers, fwx.Describe()[0])
-		}
-	}
 	var customDns *backends.InstanceDNS
 	if tags[TAG_DNS_DOMAIN_NAME] != "" {
 		customDns = &backends.InstanceDNS{
@@ -193,9 +186,9 @@ func (s *b) getInstanceDetails(log *logger.Logger, inst *computepb.Instance, vol
 		CustomDNS: customDns,
 		BackendSpecific: &InstanceDetail{
 			Volumes:          volslist,
-			FirewallList:     fwPointers,
-			Network:          net,
-			Subnet:           sub,
+			FirewallIDs:      firewalls,
+			NetworkID:        net.NetworkId,
+			SubnetID:         sub.SubnetId,
 			LabelFingerprint: inst.GetLabelFingerprint(),
 			FirewallTags:     inst.GetTags().GetItems(),
 			TagFingerprint:   inst.GetTags().GetFingerprint(),
@@ -476,7 +469,7 @@ func (s *b) InstancesTerminate(instances backends.InstanceList, waitDur time.Dur
 	}
 
 	// if no more instances exist for this project, delete the ssh key from amazon and locally from filepath.Join(s.sshKeysDir, s.project)
-	if removeSSHKey {
+	if removeSSHKey && s.createInstanceCount.Get() == 0 {
 		log.Detail("Remove SSH keys as no more instances exist for this project")
 		os.Remove(filepath.Join(s.sshKeysDir, s.project))
 		os.Remove(filepath.Join(s.sshKeysDir, s.project+".pub"))
@@ -1016,6 +1009,8 @@ func (s *b) CreateInstances(input *backends.CreateInstanceInput, waitDur time.Du
 	log := s.log.WithPrefix("CreateInstances: job=" + shortuuid.New() + " ")
 	log.Detail("Start")
 	defer log.Detail("End")
+	s.createInstanceCount.Inc()
+	defer s.createInstanceCount.Dec()
 
 	// early check - DNS
 	if input.CustomDNS != nil && input.CustomDNS.Name != "" && input.Nodes > 1 {
